A messy bug fix
A very long time to figure it was my healing hearts where the porblem, whilst on build i notice some of my triggers where not working but randomly and not with any sense until i relised my players where getting removed. I noticed this and only think i thought was my healing hearts as thats the only collide i said to get rid to stop mulitple and it was the case.