Got to demo my game for others
I got useful feedback which I added what I could, I made a tutorial so players knew what to do, made the starting area a bit easier(difficulty might be an issue). Changed the colour of the breakable wood that would make it easier. And the advice form Xander form the C++ lab was try to add particle effects and fog.